+++ ErsatzTV.Application/Configuration/Commands/UpdateGeneralSettings.cs
@@ -1,0 +1,5 @@
+﻿using ErsatzTV.Core;
+
+namespace ErsatzTV.Application.Configuration;
+
+public record UpdateGeneralSettings(GeneralSettingsViewModel GeneralSettings) : IRequest<Either<BaseError, Unit>>;