+++ ErsatzTV.Application/Configuration/GeneralSettingsViewModel.cs
@@ -1,0 +1,8 @@
+using Serilog.Events;
+
+namespace ErsatzTV.Application.Configuration;
+
+public class GeneralSettingsViewModel
+{
+    public LogEventLevel MinimumLogLevel { get; set; }
+}