+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - In previous versions, some libraries would incorrectly display only one item
 - Properly display old versions of renamed items in trash
 
+### Added
+- Add `Minimum Log Level` option to `Settings` page
+  - Other methods of configuring the log level will no longer work
+
 ## [0.6.2-beta] - 2022-06-18
 ### Fixed
 - Fix content repeating for up to a minute near the top of every hour